requirements.txt
Flask
llama-index
llama-index-llms-ollama
llama-index-llms-replicate
llama-index-embeddings-huggingface
llama-parse
Werkzeug==2.2.2
python-dotenv
arize-phoenix
openinference-instrumentation-llama-index
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-proto>=1.12.0
llama-index-callbacks-arize-phoenix
